Move pages into subdir
