+++ pt_os_web_portal/miniscreen_onboarding/menu_pages/__init__.py
@@ -1,4 +1,4 @@
-from .ap import ApMenuPage
-from .carry_on import CarryOnMenuPage
-from .open_browser import OpenBrowserMenuPage
-from .welcome import WelcomeMenuPage
+from .pages.ap import ApMenuPage
+from .pages.carry_on import CarryOnMenuPage
+from .pages.open_browser import OpenBrowserMenuPage
+from .pages.welcome import WelcomeMenuPage
