+++ pt_os_web_portal/miniscreen_onboarding/menu_pages/pages/welcome.py
